Prefer retail UI assets; add msbuild project targets
Resolve developer workflow and parity tooling changes:

- .vscode/build.ps1: add explicitProjectTargetMap and logic to resolve /t:Targets to specific vcxproj paths; introduce per-project MSBuild args and build projects directly when all requested targets map to known projects. Simplify UI bundle handling by removing Python-driven pak generation and instead remove stale local PK3s; log that retail UI assets are loaded from the Quake Live installation.
- .vscode/launch.json / .vscode/launch.ps1: point fs_cdpath to the Steam Quake Live install and make the launch script use the retail install as the UI root (remove staging-root checks and runtime bundle refresh). Also remove runtime generation/refresh of pak_uiql.pk3 from launch path and delete stale PK3s on startup.
- IMPLEMENTATION_PLAN.md: bump last-updated date and add Task A182 documenting completion of feeding the retail arena catalog into the Awesomium map list and related parity work.
- artifacts/ui_bundle/runtime_ui_package_manifest.json: update main_package size and sha256.
- Add new docs/plans/2026-06-04-networking.md: lengthy netcode retail-parity plan and roadmap (netcode parity manifest, corpus, harness, CI, etc.).
- Add tests/test_netcode_parity_manifest.py and other test/tool updates to support the parity efforts.

Also include assorted source, test, and tooling updates related to Awesomium/web UI parity and networking (see modified files under src/, tests/, scripts/, and tools/). This change centralizes UI asset sourcing to the retail installation and improves targeted MSBuild project builds and parity tooling.

+++ src/game/g_config.c
@@ -731,7 +731,7 @@ uint64_t G_ComputeConfigCustomSettingsMask( void ) {
 		 g_weaponConfig.grenadeSplashDamage != 100 ||
 		 g_weaponConfig.grenadeSplashRadius != 150 ||
 		 g_weaponConfig.grenadeSpeed != 700 ||
-		 G_ConfigFloatDiffersFromDefault( g_knockbackConfig.grenadeLauncher, DEFAULT_KNOCKBACK_GL ) ) {
+		 G_ConfigFloatDiffersFromDefault( g_knockbackConfig.grenadeLauncher, (float)DEFAULT_KNOCKBACK_GL ) ) {
 		mask |= CUSTOM_SETTING_GRENADE_LAUNCHER;
 	}
 
@@ -741,8 +741,8 @@ uint64_t G_ComputeConfigCustomSettingsMask( void ) {
 		 g_weaponConfig.rocketSplashRadius != 120 ||
 		 g_weaponConfig.rocketSpeed != 1000 ||
 		 g_weaponConfig.rocketSplashOffset != -10 ||
-		 G_ConfigFloatDiffersFromDefault( g_knockbackConfig.rocketLauncher, DEFAULT_KNOCKBACK_RL ) ||
-		 G_ConfigFloatDiffersFromDefault( g_knockbackConfig.rocketLauncherSelf, DEFAULT_KNOCKBACK_RL_SELF ) ||
+		 G_ConfigFloatDiffersFromDefault( g_knockbackConfig.rocketLauncher, (float)DEFAULT_KNOCKBACK_RL ) ||
+		 G_ConfigFloatDiffersFromDefault( g_knockbackConfig.rocketLauncherSelf, (float)DEFAULT_KNOCKBACK_RL_SELF ) ||
 		 G_ConfigFloatDiffersFromDefault( g_weaponConfig.rocketAccelerationFactor, 1.0f ) ||
 		 g_weaponConfig.rocketAccelerationRate != 16 ) {
 		mask |= CUSTOM_SETTING_ROCKET_LAUNCHER;
@@ -758,7 +758,7 @@ uint64_t G_ComputeConfigCustomSettingsMask( void ) {
 
 	if ( g_weaponReloadConfig.railgun != DEFAULT_WEAPON_RELOAD_RG ||
 		 g_weaponConfig.railgunDamage != 80 ||
-		 G_ConfigFloatDiffersFromDefault( g_knockbackConfig.railgun, DEFAULT_KNOCKBACK_RG ) ) {
+		 G_ConfigFloatDiffersFromDefault( g_knockbackConfig.railgun, (float)DEFAULT_KNOCKBACK_RG ) ) {
 		mask |= CUSTOM_SETTING_RAILGUN;
 	}
 
@@ -767,8 +767,8 @@ uint64_t G_ComputeConfigCustomSettingsMask( void ) {
 		 g_weaponConfig.plasmaSplashDamage != 15 ||
 		 g_weaponConfig.plasmaSplashRadius != 20 ||
 		 g_weaponConfig.plasmaSpeed != 2000 ||
-		 G_ConfigFloatDiffersFromDefault( g_knockbackConfig.plasmagun, DEFAULT_KNOCKBACK_PG ) ||
-		 G_ConfigFloatDiffersFromDefault( g_knockbackConfig.plasmagunSelf, DEFAULT_KNOCKBACK_PG_SELF ) ||
+		 G_ConfigFloatDiffersFromDefault( g_knockbackConfig.plasmagun, (float)DEFAULT_KNOCKBACK_PG ) ||
+		 G_ConfigFloatDiffersFromDefault( g_knockbackConfig.plasmagunSelf, (float)DEFAULT_KNOCKBACK_PG_SELF ) ||
 		 G_ConfigFloatDiffersFromDefault( g_weaponConfig.plasmaAccelerationFactor, 1.0f ) ||
 		 g_weaponConfig.plasmaAccelerationRate != 16 ) {
 		mask |= CUSTOM_SETTING_PLASMAGUN;
@@ -901,13 +901,13 @@ void G_InitKnockbackConfig( void ) {
 	g_knockbackConfig.gauntlet = G_ReadKnockbackCvar( &g_knockback_g, DEFAULT_KNOCKBACK_G, "g_knockback_g" );
 	g_knockbackConfig.machinegun = G_ReadKnockbackCvar( &g_knockback_mg, DEFAULT_KNOCKBACK_MG, "g_knockback_mg" );
 	g_knockbackConfig.shotgun = G_ReadKnockbackCvar( &g_knockback_sg, DEFAULT_KNOCKBACK_SG, "g_knockback_sg" );
-	g_knockbackConfig.grenadeLauncher = G_ReadKnockbackCvar( &g_knockback_gl, DEFAULT_KNOCKBACK_GL, "g_knockback_gl" );
-	g_knockbackConfig.rocketLauncher = G_ReadKnockbackCvar( &g_knockback_rl, DEFAULT_KNOCKBACK_RL, "g_knockback_rl" );
-	g_knockbackConfig.rocketLauncherSelf = G_ReadKnockbackCvar( &g_knockback_rl_self, DEFAULT_KNOCKBACK_RL_SELF, "g_knockback_rl_self" );
+	g_knockbackConfig.grenadeLauncher = G_ReadKnockbackCvar( &g_knockback_gl, (float)DEFAULT_KNOCKBACK_GL, "g_knockback_gl" );
+	g_knockbackConfig.rocketLauncher = G_ReadKnockbackCvar( &g_knockback_rl, (float)DEFAULT_KNOCKBACK_RL, "g_knockback_rl" );
+	g_knockbackConfig.rocketLauncherSelf = G_ReadKnockbackCvar( &g_knockback_rl_self, (float)DEFAULT_KNOCKBACK_RL_SELF, "g_knockback_rl_self" );
 	g_knockbackConfig.lightningGun = G_ReadKnockbackCvar( &g_knockback_lg, DEFAULT_KNOCKBACK_LG, "g_knockback_lg" );
-	g_knockbackConfig.railgun = G_ReadKnockbackCvar( &g_knockback_rg, DEFAULT_KNOCKBACK_RG, "g_knockback_rg" );
-	g_knockbackConfig.plasmagun = G_ReadKnockbackCvar( &g_knockback_pg, DEFAULT_KNOCKBACK_PG, "g_knockback_pg" );
-	g_knockbackConfig.plasmagunSelf = G_ReadKnockbackCvar( &g_knockback_pg_self, DEFAULT_KNOCKBACK_PG_SELF, "g_knockback_pg_self" );
+	g_knockbackConfig.railgun = G_ReadKnockbackCvar( &g_knockback_rg, (float)DEFAULT_KNOCKBACK_RG, "g_knockback_rg" );
+	g_knockbackConfig.plasmagun = G_ReadKnockbackCvar( &g_knockback_pg, (float)DEFAULT_KNOCKBACK_PG, "g_knockback_pg" );
+	g_knockbackConfig.plasmagunSelf = G_ReadKnockbackCvar( &g_knockback_pg_self, (float)DEFAULT_KNOCKBACK_PG_SELF, "g_knockback_pg_self" );
 	g_knockbackConfig.bfg = G_ReadKnockbackCvar( &g_knockback_bfg, DEFAULT_KNOCKBACK_BFG, "g_knockback_bfg" );
 	g_knockbackConfig.grapplingHook = G_ReadKnockbackCvar( &g_knockback_gh, DEFAULT_KNOCKBACK_GH, "g_knockback_gh" );
 	g_knockbackConfig.nailgun = G_ReadKnockbackCvar( &g_knockback_ng, DEFAULT_KNOCKBACK_NG, "g_knockback_ng" );
